apps/api/requirements.txt
# ──────────────────────────────────────────────────────────────
# DesignAI API — Python dependencies
# Install: pip install -r requirements.txt
# ──────────────────────────────────────────────────────────────

# Web framework
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20

# Data validation
pydantic==2.10.3
pydantic-settings==2.7.0

# AI
anthropic==0.44.0

# HTTP client
httpx==0.28.1
PyJWT==2.10.1
python-dateutil==2.9.0

# Environment
python-dotenv==1.0.1

# Supabase
supabase==2.11.0

# Geospatial
shapely==2.0.6
geopandas==1.0.1
pyproj==3.7.0

# 3D geometry / mesh
trimesh==4.5.3

# CAD export
ezdxf==1.3.4

# PDF generation
weasyprint==63.1
jinja2==3.1.5

# Task queue
celery==5.4.0
redis==5.2.1

# Utilities
orjson==3.10.12
